refactor(tui): isolate Zed editor context polling
